% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_string.R
\name{combine_string}
\alias{combine_string}
\title{combine_string}
\usage{
combine_string(
  libraries = "library(ggplot2)",
  init_layer,
  geoms,
  facet = NULL,
  labs = NULL,
  theme_std = NULL,
  theme_custom = NULL,
  palette_fill = NULL,
  palette_color = NULL
)
}
\value{
unk
}
\description{

}
\references{

}
\author{
K. Enevoldsen
}
