% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_string.R
\name{create_geom}
\alias{create_geom}
\title{create_geom}
\usage{
create_geom(geom, color = NULL, fill = NULL, shape = NULL, alpha = NULL)
}
\value{
unk
}
\description{

}
\references{

}
\author{
K. Enevoldsen
}
