% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_string.R
\name{create_init}
\alias{create_init}
\title{create_init}
\usage{
create_init(
  df = "dataset",
  x,
  y = NULL,
  fill = NULL,
  color = NULL,
  shape = NULL
)
}
\value{
unk
}
\description{

}
\references{

}
\author{
K. Enevoldsen
}
