% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_string.R
\name{create_labs}
\alias{create_labs}
\title{create_labs}
\usage{
create_labs(
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  tag = NULL,
  color = NULL,
  fill = NULL,
  shape = NULL,
  x = NULL,
  y = NULL
)
}
\value{
unk
}
\description{

}
\references{

}
\author{
K. Enevoldsen
}
