% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_string.R
\name{create_stat_geom}
\alias{create_stat_geom}
\title{create_stat_geom}
\usage{
create_stat_geom(type = "overlay_norm", var = NULL)
}
\value{
unk
}
\description{

}
\references{

}
\author{
K. Enevoldsen
}
