% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_string.R
\name{geom_replace_arg}
\alias{geom_replace_arg}
\title{geom_replace_arg}
\usage{
geom_replace_arg(str, arg_to_rep = "x", replacement)
}
\value{
unk
}
\description{

}
\references{

}
\author{
K. Enevoldsen
}
