% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{gglearn}
\alias{gglearn}
\title{gglearn: Shiny App for learning ggplot2}
\usage{
gglearn(dataset)
}
\arguments{
\item{df}{A dataframe object}
}
\value{
start a shiny application
}
\description{
A shiny application for teaching and learning ggplot2
}
\references{

}
\author{
L. Hansen
}
