% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R, R/gglearn2.R
\docType{package}
\name{gglearn2}
\alias{gglearn2}
\title{gglearn2: Shiny App for learning ggplot2}
\usage{
gglearn2(df = iris)
}
\arguments{
\item{df}{A dataframe object}
}
\value{
start a shiny application
}
\description{
A shiny application for teaching and learning ggplot2

A package for teaching ggplot2
}
\references{

}
\author{
L. Hansen

L. Hansen \email{lasseh0310@gmail.com} and K. Enevoldsen \email{kennethcenevoldsen@gmail.com}
}
