% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglearn2.R
\docType{package}
\name{gglearn2}
\alias{gglearn2}
\title{gglearn2}
\description{
A package for teaching ggplot2
}
\author{
L. Hansen \email{lasseh0310@gmail.com} and K. Enevoldsen \email{kennethcenevoldsen@gmail.com}
}
